individual-parts/teamlead/ready-code/requirements.txt
Django==5.1.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
django-cors-headers==4.6.0
django-filter==24.3
argon2-cffi==23.1.0
python-dotenv==1.0.1
Pillow==11.1.0
